Replace calls to mj_clone in lib.sjs
The calls to mj_clone in lib.sjs (e.g. in mc_add_SM) have been replaced
with calls to mj_new_from() which creates a zeroed array with the same
shape and size as the array given in input.  This is done so that calls
to mj_clone() can be tallied in comparing the number of copies between
different implementations.